setup the the web app to render a 401 web page
